enable the test suite
The test suite should now succeed reliably, given the prerequisites.
Enable it by default on Fedora and RHEL8+ and allow to build with or
without using the tests conditionally.

On RHEL, we run the test suite with `--full-scan` do avoid a Heisenbug
which shows at least on some archictectures (s390x, aarch) with some
file systems (xfs, tmpfs) in some infrastructure (koji scratch, copr).
It is not a runtime bug: At worst, `notmuch new` misses a new message on
one run and picks it up on the next run. In the test suite though, this
can lead to spurious fails because of wrong message counts (-1 on
subtest, +1 on the next).