add explicit dependency of bindings on notmuch nvr
For most bindings, rpm picks up the implicit dependency on the devel .so
but not on the versioned .so nor on the exact notmuch version. So far,
this has not created any problems besides /rpmdeps erros from
rpminspect.

Be proactive and make the bindings depend on notmuch nvr explicitely.
(Other subpackages do so already.)